+++ aws-java-sdk-ec2/src/main/java/com/amazonaws/services/ec2/model/ElasticInferenceAccelerator.java
@@ -28,21 +28,28 @@ public class ElasticInferenceAccelerator implements Serializable, Cloneable {
 
     /**
      * <p>
-     * The type of elastic inference accelerator. The possible values are <code>eia1.small</code>,
-     * <code>eia1.medium</code>, and <code>eia1.large</code>.
+     * The type of elastic inference accelerator. The possible values are <code>eia1.medium</code>,
+     * <code>eia1.large</code>, and <code>eia1.xlarge</code>.
      * </p>
      */
     private String type;
+    /**
+     * <p>
+     * The number of elastic inference accelerators of given type to be attached to the instance. Only positive values
+     * allowed. If not specified defaults to 1.
+     * </p>
+     */
+    private Integer count;
 
     /**
      * <p>
-     * The type of elastic inference accelerator. The possible values are <code>eia1.small</code>,
-     * <code>eia1.medium</code>, and <code>eia1.large</code>.
+     * The type of elastic inference accelerator. The possible values are <code>eia1.medium</code>,
+     * <code>eia1.large</code>, and <code>eia1.xlarge</code>.
      * </p>
      * 
      * @param type
-     *        The type of elastic inference accelerator. The possible values are <code>eia1.small</code>,
-     *        <code>eia1.medium</code>, and <code>eia1.large</code>.
+     *        The type of elastic inference accelerator. The possible values are <code>eia1.medium</code>,
+     *        <code>eia1.large</code>, and <code>eia1.xlarge</code>.
      */
 
     public void setType(String type) {
@@ -51,12 +58,12 @@ public class ElasticInferenceAccelerator implements Serializable, Cloneable {
 
     /**
      * <p>
-     * The type of elastic inference accelerator. The possible values are <code>eia1.small</code>,
-     * <code>eia1.medium</code>, and <code>eia1.large</code>.
+     * The type of elastic inference accelerator. The possible values are <code>eia1.medium</code>,
+     * <code>eia1.large</code>, and <code>eia1.xlarge</code>.
      * </p>
      * 
-     * @return The type of elastic inference accelerator. The possible values are <code>eia1.small</code>,
-     *         <code>eia1.medium</code>, and <code>eia1.large</code>.
+     * @return The type of elastic inference accelerator. The possible values are <code>eia1.medium</code>,
+     *         <code>eia1.large</code>, and <code>eia1.xlarge</code>.
      */
 
     public String getType() {
@@ -65,18 +72,64 @@ public class ElasticInferenceAccelerator implements Serializable, Cloneable {
 
     /**
      * <p>
-     * The type of elastic inference accelerator. The possible values are <code>eia1.small</code>,
-     * <code>eia1.medium</code>, and <code>eia1.large</code>.
+     * The type of elastic inference accelerator. The possible values are <code>eia1.medium</code>,
+     * <code>eia1.large</code>, and <code>eia1.xlarge</code>.
      * </p>
      * 
      * @param type
-     *        The type of elastic inference accelerator. The possible values are <code>eia1.small</code>,
-     *        <code>eia1.medium</code>, and <code>eia1.large</code>.
+     *        The type of elastic inference accelerator. The possible values are <code>eia1.medium</code>,
+     *        <code>eia1.large</code>, and <code>eia1.xlarge</code>.
      * @return Returns a reference to this object so that method calls can be chained together.
      */
 
     public ElasticInferenceAccelerator withType(String type) {
         setType(type);
+        return this;
+    }
+
+    /**
+     * <p>
+     * The number of elastic inference accelerators of given type to be attached to the instance. Only positive values
+     * allowed. If not specified defaults to 1.
+     * </p>
+     * 
+     * @param count
+     *        The number of elastic inference accelerators of given type to be attached to the instance. Only positive
+     *        values allowed. If not specified defaults to 1.
+     */
+
+    public void setCount(Integer count) {
+        this.count = count;
+    }
+
+    /**
+     * <p>
+     * The number of elastic inference accelerators of given type to be attached to the instance. Only positive values
+     * allowed. If not specified defaults to 1.
+     * </p>
+     * 
+     * @return The number of elastic inference accelerators of given type to be attached to the instance. Only positive
+     *         values allowed. If not specified defaults to 1.
+     */
+
+    public Integer getCount() {
+        return this.count;
+    }
+
+    /**
+     * <p>
+     * The number of elastic inference accelerators of given type to be attached to the instance. Only positive values
+     * allowed. If not specified defaults to 1.
+     * </p>
+     * 
+     * @param count
+     *        The number of elastic inference accelerators of given type to be attached to the instance. Only positive
+     *        values allowed. If not specified defaults to 1.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     */
+
+    public ElasticInferenceAccelerator withCount(Integer count) {
+        setCount(count);
         return this;
     }
 
@@ -93,7 +146,9 @@ public class ElasticInferenceAccelerator implements Serializable, Cloneable {
         StringBuilder sb = new StringBuilder();
         sb.append("{");
         if (getType() != null)
-            sb.append("Type: ").append(getType());
+            sb.append("Type: ").append(getType()).append(",");
+        if (getCount() != null)
+            sb.append("Count: ").append(getCount());
         sb.append("}");
         return sb.toString();
     }
@@ -112,6 +167,10 @@ public class ElasticInferenceAccelerator implements Serializable, Cloneable {
             return false;
         if (other.getType() != null && other.getType().equals(this.getType()) == false)
             return false;
+        if (other.getCount() == null ^ this.getCount() == null)
+            return false;
+        if (other.getCount() != null && other.getCount().equals(this.getCount()) == false)
+            return false;
         return true;
     }
 
@@ -121,6 +180,7 @@ public class ElasticInferenceAccelerator implements Serializable, Cloneable {
         int hashCode = 1;
 
         hashCode = prime * hashCode + ((getType() == null) ? 0 : getType().hashCode());
+        hashCode = prime * hashCode + ((getCount() == null) ? 0 : getCount().hashCode());
         return hashCode;
     }