+++ aws-java-sdk-ec2/src/main/java/com/amazonaws/services/ec2/model/transform/CreateLaunchTemplateRequestMarshaller.java
@@ -386,6 +386,11 @@ public class CreateLaunchTemplateRequestMarshaller implements Marshaller<Request
                         request.addParameter("LaunchTemplateData.ElasticInferenceAccelerator." + elasticInferenceAcceleratorsListIndex + ".Type",
                                 StringUtils.fromString(requestLaunchTemplateDataElasticInferenceAcceleratorsListValue.getType()));
                     }
+
+                    if (requestLaunchTemplateDataElasticInferenceAcceleratorsListValue.getCount() != null) {
+                        request.addParameter("LaunchTemplateData.ElasticInferenceAccelerator." + elasticInferenceAcceleratorsListIndex + ".Count",
+                                StringUtils.fromInteger(requestLaunchTemplateDataElasticInferenceAcceleratorsListValue.getCount()));
+                    }
                     elasticInferenceAcceleratorsListIndex++;
                 }
             }