+++ aws-java-sdk-ec2/src/main/java/com/amazonaws/services/ec2/model/transform/LaunchTemplateElasticInferenceAcceleratorResponseStaxUnmarshaller.java
@@ -48,6 +48,11 @@ public class LaunchTemplateElasticInferenceAcceleratorResponseStaxUnmarshaller i
                     launchTemplateElasticInferenceAcceleratorResponse.setType(StringStaxUnmarshaller.getInstance().unmarshall(context));
                     continue;
                 }
+
+                if (context.testExpression("count", targetDepth)) {
+                    launchTemplateElasticInferenceAcceleratorResponse.setCount(IntegerStaxUnmarshaller.getInstance().unmarshall(context));
+                    continue;
+                }
             } else if (xmlEvent.isEndElement()) {
                 if (context.getCurrentDepth() < originalDepth) {
                     return launchTemplateElasticInferenceAcceleratorResponse;