+++ aws-java-sdk-ec2/src/main/java/com/amazonaws/services/ec2/model/transform/LaunchTemplateElasticInferenceAcceleratorStaxUnmarshaller.java
@@ -48,6 +48,11 @@ public class LaunchTemplateElasticInferenceAcceleratorStaxUnmarshaller implement
                     launchTemplateElasticInferenceAccelerator.setType(StringStaxUnmarshaller.getInstance().unmarshall(context));
                     continue;
                 }
+
+                if (context.testExpression("Count", targetDepth)) {
+                    launchTemplateElasticInferenceAccelerator.setCount(IntegerStaxUnmarshaller.getInstance().unmarshall(context));
+                    continue;
+                }
             } else if (xmlEvent.isEndElement()) {
                 if (context.getCurrentDepth() < originalDepth) {
                     return launchTemplateElasticInferenceAccelerator;